Migrate OverrideDatastoreController to override_datastore blueprint
