+++ ckanext/datagovmk/templates/package/resources.html
@@ -28,7 +28,7 @@ along with this program.  If not, see <https://www.gnu.org/licenses/>.
       {% endfor %}
     </ul>
   {% else %}
-    {% trans url=h.url_for(controller='package', action='new_resource', id=pkg.name) %}
+    {% trans url=h.url_for(controller='dataset', action='new_resource', id=pkg.name) %}
       <p class="empty">This dataset has no data, <a href="{{ url }}">why not add some?</a></p>
     {% endtrans %}
   {% endif %}
