+++ ckanext/datagovmk/views/__init__.py
@@ -1,1 +1,3 @@
 from ckanext.datagovmk.views.bulk_download import bulk_download
+from ckanext.datagovmk.views.report_issue import report_issue
+from ckanext.datagovmk.views.override_datastore import override_datastore
